refactor: extract settings sections into sub-pages
Move credentials, notifications, and appearance into dedicated
sub-pages to keep the settings screen compact. Add platform-aware
styling (glass on iOS, material on Android) to the appearance page.
Guide users to iOS Settings when "Always Allow" location is needed.
