+++ src/app/app/_layout.tsx
@@ -87,6 +87,10 @@ function AppContent() {
     <DialogProvider>
       <Stack>
         <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+        <Stack.Screen name="kantine-login" options={{ headerShown: false, presentation: 'card' }} />
+        <Stack.Screen name="automaten-login" options={{ headerShown: false, presentation: 'card' }} />
+        <Stack.Screen name="notifications" options={{ headerShown: false, presentation: 'card' }} />
+        <Stack.Screen name="appearance" options={{ headerShown: false, presentation: 'card' }} />
       </Stack>
       <StatusBar style="auto" />
     </DialogProvider>
